requirements.txt
asgiref==3.8.1
boto3==1.35.96
botocore==1.35.96
Django==4.2.6
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
git-filter-repo==2.47.0
greenlet==3.1.1
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==1.4.54
sqlalchemy-redshift==0.8.14
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
django-redis==5.4.0
redis==5.0.1
django-environ==0.10.0
pyfcm==1.5.0